# contracts/openapi/core-api.v1.yaml
openapi: 3.1.0
info:
  title: GDC Core API
  version: 1.0.0
  description: API principal del backend GDC 2.0 (.NET 10 + EF Core + PostgreSQL)
servers:
  - url: http://localhost:40203
    description: Desarrollo local
paths:
...
  /api/v1/notif/rules:
    get:
      operationId: listNotifRules
      summary: Lista reglas de comunicación del tenant
      tags: [NOTIF Rules]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reglas configuradas
    post:
      operationId: createNotifRule
      summary: Crea regla cronológica o estatal
      tags: [NOTIF Rules]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotifRuleRequest"
      responses:
        "201":
          description: Regla creada
        "400":
          description: Validación fallida
  /api/v1/notif/rules/{id}:
    get:
      operationId: getNotifRule
      summary: Obtiene regla por id
      tags: [NOTIF Rules]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Regla encontrada
        "404":
          description: No encontrada
    put:
      operationId: updateNotifRule
      summary: Actualiza regla
      tags: [NOTIF Rules]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotifRuleRequest"
      responses:
        "200":
          description: Regla actualizada
        "400":
          description: Validación fallida
        "404":
          description: No encontrada
    delete:
      operationId: deleteNotifRule
      summary: Elimina regla (soft delete)
      tags: [NOTIF Rules]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Eliminada
        "404":
          description: No encontrada
...
  /api/v1/reglas/runs:
    get:
      operationId: listReglasRuns
      summary: Lista corridas de evaluación REGLAS
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Corridas del tenant
    post:
      operationId: triggerReglasRun
      summary: Dispara evaluación manual de reglas activas
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "202":
          description: Corrida aceptada
        "400":
          description: Error en evaluación
  /api/v1/reglas/process:
    post:
      operationId: processReglasPendingMatches
      summary: Orquesta PDF y correo para coincidencias pendientes
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Resultado de procesamiento
  /api/v1/reglas/runs/{id}:
    get:
      operationId: getReglasRun
      summary: Obtiene corrida de evaluación por id
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Corrida encontrada
        "404":
          description: No encontrada
    post:
      operationId: processReglasRun
      summary: Orquesta PDF y correo para coincidencias de una corrida
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Resultado de procesamiento
        "404":
          description: Corrida no encontrada
  /api/v1/reglas/matches:
    get:
      operationId: listReglasMatches
      summary: Lista coincidencias regla-comparendo
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: runId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: ruleId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Coincidencias registradas
...
  /api/v1/reglas/rules:
    get:
      operationId: listReglasRules
      summary: Lista reglas dinámicas DP del tenant
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Reglas configuradas
    post:
      operationId: createReglasRule
      summary: Crea regla dinámica con árbol de condiciones
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReglasRuleRequest"
      responses:
        "201":
          description: Regla creada
        "400":
          description: Validación fallida
  /api/v1/reglas/rules/{id}:
    get:
      operationId: getReglasRule
      summary: Obtiene regla dinámica por id
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Regla encontrada
        "404":
          description: No encontrada
    put:
      operationId: updateReglasRule
      summary: Actualiza regla dinámica y condiciones
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReglasRuleRequest"
      responses:
        "200":
          description: Regla actualizada
        "400":
          description: Validación fallida
        "404":
          description: No encontrada
    delete:
      operationId: deleteReglasRule
      summary: Elimina regla dinámica (soft delete)
      tags: [Reglas]
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Eliminada
        "404":
          description: No encontrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateNotifRuleRequest:
      type: object
      required: [emailTemplateId, name, triggerType]
      properties:
        emailTemplateId:
          type: string
          format: uuid
        name:
          type: string
        triggerType:
          type: string
          enum: [chronological, state]
        triggerDays:
          type: integer
          nullable: true
        triggerReference:
          type: string
          enum: [fecha_comparendo, fecha_notificacion]
          nullable: true
        triggerEstado:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
    UpdateNotifRuleRequest:
      allOf:
        - $ref: "#/components/schemas/CreateNotifRuleRequest"
...
    ReglasConditionNode:
      type: object
      required: [nodeType]
      properties:
        nodeType:
          type: string
          enum: [group, predicate]
        logicOperator:
          type: string
          enum: [and, or]
          nullable: true
        fieldKey:
          type: string
          enum: [estado, numero_comparendo, placa, infractor_nombre, fecha_comparendo, fecha_notificacion, destino_email]
          nullable: true
        comparisonOperator:
          type: string
          enum: [eq, neq, gt, lt, contains, not_contains]
          nullable: true
        comparisonValue:
          type: string
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/ReglasConditionNode"
          nullable: true
...
    CreateReglasRuleRequest:
      type: object
      required: [name, isActive, pdfTemplateId, emailSubject, emailBodyHtml]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        pdfTemplateId:
          type: string
          format: uuid
        emailSubject:
          type: string
        emailBodyHtml:
          type: string
        secretariatContactId:
          type: string
          format: uuid
          nullable: true
        conditionRoot:
          $ref: "#/components/schemas/ReglasConditionNode"
          nullable: true
    UpdateReglasRuleRequest:
      allOf:
        - $ref: "#/components/schemas/CreateReglasRuleRequest"
...